refactor: introduce directory for processors
