+++ src/core/encryption/ciphers/AES/index.ts
@@ -1,15 +1,15 @@
-import { IntegrityError } from '../../EncryptionIntegrityCheck';
+import { IntegrityError } from '../../processors/BufferIntegrityProcessor';
 import { joinBuffers } from '../../utils/buffers';
 import { getRandomBytes } from '../../utils/random';
 
-import { ICipher } from '../..';
+import { IEncryptionProcessor } from '../..';
 
 /**
  * AES-GCM cipher
  * MDN: https://developer.mozilla.org/en-US/docs/Web/API/AesGcmParams
  * Algorithm recommendations: https://nvlpubs.nist.gov/nistpubs/Legacy/SP/nistspecialpublication800-38d.pdf
  */
-export class AESGCMCipher implements ICipher {
+export class AESGCMCipher implements IEncryptionProcessor {
 	private readonly ivSize = 96;
 
 	private readonly key;
