+++ src/core/workers/Cryptography.worker.ts
@@ -1,12 +1,12 @@
 import { WorkerMessenger } from '../../utils/workers/WorkerMessenger';
 import { WorkerRPC } from '../../utils/workers/WorkerRPC';
 
-import { BufferSizeObfuscator } from '../encryption/BufferSizeObfuscator';
 import { AESGCMCipher } from '../encryption/ciphers/AES';
-import { CascadeCipher } from '../encryption/ciphers/CascadeCipher';
 import { TwofishCTRCipher } from '../encryption/ciphers/Twofish';
 import { EncryptionController } from '../encryption/EncryptionController';
-import { EncryptionIntegrityCheck } from '../encryption/EncryptionIntegrityCheck';
+import { BufferIntegrityProcessor } from '../encryption/processors/BufferIntegrityProcessor';
+import { BufferSizeObfuscationProcessor } from '../encryption/processors/BufferSizeObfuscationProcessor';
+import { CascadeCipherProcessor } from '../encryption/processors/CascadeCipherProcessor';
 import { getDerivedKeysManager, getMasterKey } from '../encryption/utils/keys';
 
 console.log('Hello world from worker');
@@ -35,9 +35,9 @@ requests.addHandler('init', async ({ secretKey, salt }) => {
 		.then((buffer) => new Uint8Array(buffer));
 
 	encryptionController = new EncryptionController(
-		new EncryptionIntegrityCheck(
-			new BufferSizeObfuscator(
-				new CascadeCipher([
+		new BufferIntegrityProcessor(
+			new BufferSizeObfuscationProcessor(
+				new CascadeCipherProcessor([
 					new AESGCMCipher(aesKey),
 					new TwofishCTRCipher(twofishKey),
 				]),
